requirements.txt
streamlit
streamlit-option-menu
streamlit-player
st-multimodal-chatinput
google-generativeai
pillow
python-dotenv